+++ storage/plugins/emeroteca/EmerotecaPlugin.php
@@ -1651,7 +1651,7 @@ class EmerotecaPlugin
         $public = 'App\\Plugins\\Emeroteca\\Controllers\\PublicController';
 
         $articles = 'App\\Plugins\\Emeroteca\\Controllers\\ContributionController';
-        foreach (['' => 'index', '/create' => 'form', '/{id:[0-9]+}' => 'form', '/import' => 'importForm', '/export' => 'export', '/{id:[0-9]+}/pdf' => 'pdf'] as $path => $method) {
+        foreach (['' => 'index', '/create' => 'form', '/{id:[0-9]+}' => 'form', '/import' => 'importForm', '/export' => 'export', '/issues' => 'issueOptions', '/{id:[0-9]+}/pdf' => 'pdf'] as $path => $method) {
             $app->get('/admin/periodicals/articles' . $path, function ($rq, $rs, $args) use ($plugin, $articles, $method) {
                 return $plugin->dispatch($articles, $method, $rq, $rs, $args);
             })->add($adminMiddleware);